Fix loss terms: pass logits and softmax where needed
Since the imported UNet model has a final softmax layer, we undo this
softmax by applying the logit function to obtain logits.

These logits are then passed directly to CrossEntropy, and also
softmaxed into the DiceLoss.

CrossEntropy now ignores the background class, because that is purely
handled by the DiceLoss.